Cập nhật ThuongHieu

+++ src/main/java/com/BaiTapLab/Service/ThuongHieuService.java
@@ -1,7 +1,5 @@
 package com.BaiTapLab.Service;
 
-import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,22 +9,32 @@ import com.BaiTapLab.Repository.ThuongHieuRepository;
 @Service
 public class ThuongHieuService {
 	@Autowired
-	private ThuongHieuRepository thRepository;
+	ThuongHieuRepository thRepository;
 	
-	public List<ThuongHieu> findAll(){
-		return thRepository.findAll();
+	
+	public ThuongHieu createThuongHieu(ThuongHieu thuonghieu) {
+		return thRepository.save(thuonghieu);
 	}
 	
-	public ThuongHieu save(ThuongHieu th) {
-		return thRepository.save(th);
+	public void deleteThuongHieu(String thuong_hieuID) throws Exception {
+	    // Kiểm tra xem thương hiệu có tồn tại không trước khi xóa
+	    if (!thRepository.existsById(thuong_hieuID)) {
+	        throw new Exception("Thương hiệu not found with ID: " + thuong_hieuID);
+	    }
+	    // Thực hiện xóa thương hiệu
+	    thRepository.deleteById(thuong_hieuID);
 	}
 	
-	public void deleteByID(String thuong_hieuID) {
-		thRepository.deleteById(thuong_hieuID);
+	public ThuongHieu findByThuongHieuID(String thuong_hieuID) {
+		return thRepository.findById(thuong_hieuID).orElseThrow();
 	}
 	
-	public ThuongHieu findByID(String thuong_hieuID) {
-		Optional<ThuongHieu> th = thRepository.findById(thuong_hieuID);
-		return th.orElseThrow(() -> new RuntimeException("Không tồn tại!"));
-	}
+	public ThuongHieu updateThuongHieu(ThuongHieu updatedThuongHieu) throws Exception {
+        // Kiểm tra xem thương hiệu có tồn tại không
+        if (!thRepository.existsById(updatedThuongHieu.getThuong_hieuID())) {
+            throw new Exception("Thương hiệu not found with ID: " + updatedThuongHieu.getThuong_hieuID());
+        }
+     // Cập nhật voucher trong cơ sở dữ liệu
+        return thRepository.save(updatedThuongHieu);
+    }
 }
